auth_service/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.2
django-cors-headers==4.3.1
django-otp==1.3.0
django-phonenumber-field==7.3.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
kombu==5.3.5
packaging==23.2
phonenumbers==8.13.31
pillow==10.2.0
channels==4.1.0
redis==5.0.4
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyotp==2.9.0
python-dateutil==2.9.0
python-decouple==3.8
python-http-client==3.3.7
pytz==2024.1
sendgrid==6.11.0
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
pika
phonenumbers==8.13.31
pillow==10.2.0
python-decouple==3.8
django-otp==1.3.0
django-phonenumber-field==7.3.0
